# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Educational Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create or update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /users/{userId}/streak:
    patch:
      operationId: updateStreak
      tags: [users]
      summary: Update user streak
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreakUpdate"
      responses:
        "200":
          description: Updated streak
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /quizzes/start:
    post:
      operationId: startQuiz
      tags: [quizzes]
      summary: Start a new quiz session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizStartInput"
      responses:
        "201":
          description: Quiz session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizSession"
  /quizzes/{sessionId}/answer:
    post:
      operationId: submitAnswer
      tags: [quizzes]
      summary: Submit an answer
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerInput"
      responses:
        "200":
          description: Answer result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerResult"
  /quizzes/{sessionId}/complete:
    post:
      operationId: completeQuiz
      tags: [quizzes]
      summary: Complete a quiz session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizCompleteInput"
      responses:
        "200":
          description: Quiz result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, displayName, role, xpAnnual, xpWeekly, streak, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [admin, student]
        xpAnnual:
          type: integer
        xpWeekly:
          type: integer
        streak:
          type: integer
        title:
          type: ["string", "null"]
        photoURL:
          type: ["string", "null"]
        createdAt:
          type: string
    UserInput:
      type: object
      required: [email, displayName, role]
      properties:
        email:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [admin, student]
        photoURL:
          type: ["string", "null"]
    UserUpdate:
      type: object
      properties:
        displayName:
          type: string
        role:
          type: string
          enum: [admin, student]
        xpAnnual:
          type: integer
        xpWeekly:
          type: integer
        streak:
          type: integer
        title:
          type: ["string", "null"]
        photoURL:
          type: ["string", "null"]
    StreakUpdate:
      type: object
      required: [streak]
      properties:
        streak:
          type: integer
...
    Question:
      type: object
      required: [id, ruleId, questionText, options, correctAnswer, createdAt]
      properties:
        id:
          type: string
        ruleId:
          type: string
        questionText:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        hint:
          type: ["string", "null"]
        createdAt:
          type: string
...
    QuizStartInput:
      type: object
      required: [userId, ruleId, questionCount]
      properties:
        userId:
          type: string
        ruleId:
          type: string
        questionCount:
          type: integer
    QuizSession:
      type: object
      required: [sessionId, userId, ruleId, questions, startedAt]
      properties:
        sessionId:
          type: string
        userId:
          type: string
        ruleId:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        startedAt:
          type: string
    AnswerInput:
      type: object
      required: [questionId, answerText, userId]
      properties:
        questionId:
          type: string
        answerText:
          type: string
        userId:
          type: string
    AnswerResult:
      type: object
      required: [correct, correctAnswer, xpAwarded]
      properties:
        correct:
          type: boolean
        correctAnswer:
          type: string
        hint:
          type: ["string", "null"]
        xpAwarded:
          type: integer
    QuizCompleteInput:
      type: object
      required: [userId, totalCorrect, totalQuestions]
      properties:
        userId:
          type: string
        totalCorrect:
          type: integer
        totalQuestions:
          type: integer
    QuizResult:
      type: object
      required: [score, xpEarned, passed]
      properties:
        score:
          type: integer
        xpEarned:
          type: integer
        passed:
          type: boolean
        newStreak:
          type: integer
...