# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Educational Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: string;
  email: string;
  displayName: string;
  role: UserRole;
  xpAnnual: number;
  xpWeekly: number;
  streak: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  photoURL?: string | null;
  createdAt: string;
}