# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Educational Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserUpdateRole } from './userUpdateRole';

export interface UserUpdate {
  displayName?: string;
  role?: UserUpdateRole;
  xpAnnual?: number;
  xpWeekly?: number;
  streak?: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  photoURL?: string | null;
}